[FIX] Serial1 is port for pins 1&2
Extract from documentation https://wiki-content.arduino.cc/en/Guide/NANOEvery

> Serial ports on the Arduino Nano Every

The USB connector of the board is connected to the pins 11 and 12 of the SAMD11. The firmware loaded on this microcontroller supports the standard features of the USB to Serial interface common to the Arduino boards, but allows high performance transfers up to 1Mbit/s. The firmware is open source and can be modified to implement new functions. The interface to the SAMD11 / USB is Serial and the the one available on pins RX0 and TX1 is Serial1.